Version Packages (#307)
This PR was opened by the [Changesets release](https://github.com/changesets/action) GitHub action. When you're ready to do a release, you can merge this and the packages will be published to npm automatically. If you're not ready to do a release yet, that's fine, whenever you add more changesets to main, this PR will be updated.


# Releases
## ai@2.1.18

### Patch Changes

-   6648b21: Add experimental client side OpenAI function calling to Svelte bindings
-   e5b983f: feat(streams): add http error handling for openai