Add replicate example and fix event termination (#405)
Fixes Replicate's non-standard SSE termination. They send an event named `done` with `data: {}` (https://replicate.com/docs/streaming). 

**The Problem**
<img width="539" alt="CleanShot 2023-08-01 at 09 29 18@2x" src="https://github.com/vercel-labs/ai/assets/4060187/9a92e64c-5448-465a-b23c-13232c8190a2">
<img width="403" alt="CleanShot 2023-08-01 at 09 29 13@2x" src="https://github.com/vercel-labs/ai/assets/4060187/c54db2db-92d2-43a8-8025-25f3974628fb">

Our current `customParser` abstraction only handles `event.data` which in the case of termination = `{}`. However, this could be part of the regular response output, and so isn't a sufficient conditional to decide termination. As a workaround, this ships an additional termination check in `AIStreamParser`. 

If we don't like this abstraction, we can refactor `customParser` to take the whole `event` and then move this logic into `ReplicateStream`.